refactor ray package
intersections needed to be removed from the ray package and moved to the shapes as the computation of normals now require the intersection. so to avoid circular dependency the intersection logic had to be moved out.
It makes more sense to have it with shapes anyway as it belongs more to them than rays.