replace internal matrix with gonum
Using the 3rd party lib increases the performance of the ray tracer tenfold.
I left the matrix package to act as the interface for the rest of the application, minimising the necessary code changes. Might come back to my solution later, but for now the focus is on the ray tracer and this performance boost is required for further work.